Upgrade Rust toolchain to 2025-09-19
Relevant upstream PRs:
- https://github.com/rust-lang/rust/pull/146664 (Clean up `ty::Dynamic`)
- https://github.com/rust-lang/rust/pull/146728 (Clippy subtree update)

Resolves: #4368

+++ kani-driver/src/call_cbmc.rs
@@ -145,27 +145,28 @@ impl KaniSession {
             .await)
         };
 
-        let verification_results = if res.is_err() {
+        if let Ok(output) = res {
+            // The timeout wasn't reached
+            Ok(VerificationResult::from(
+                output.unwrap(),
+                harness.attributes.should_panic,
+                start_time,
+            ))
+        } else {
             // An error occurs if the timeout was reached
 
             // Kill the process
             cbmc_process.kill().await?;
 
-            VerificationResult {
+            Ok(VerificationResult {
                 status: VerificationStatus::Failure,
                 failed_properties: FailedProperties::None,
                 results: Err(ExitStatus::Timeout),
                 runtime: start_time.elapsed(),
                 generated_concrete_test: false,
                 coverage_results: None,
-            }
-        } else {
-            // The timeout wasn't reached
-            let output = res.unwrap()?;
-            VerificationResult::from(output, harness.attributes.should_panic, start_time)
-        };
-
-        Ok(verification_results)
+            })
+        }
     }
 
     /// "Internal," but also used by call_cbmc_viewer
